# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: api-tracker
  version: "1.0"
  description: API-tracker REST API. Base URL — https://api.apitracker.ru
servers:
  - url: https://api.apitracker.ru
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT access token (RS256, 15 min) or PAT (prefix `pat_`).
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, description, owner_id, created_at, updated_at]
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        owner_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, description, owner_id, created_at, updated_at]
...
    ProjectOwnershipTransfer:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        from_user_id: { type: string, format: uuid }
        to_user_id: { type: string, format: uuid }
        status: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, project_id, from_user_id, to_user_id, status, created_at]
    TeamOwnershipTransfer:
      type: object
      properties:
        id: { type: string, format: uuid }
        team_id: { type: string, format: uuid }
        from_user_id: { type: string, format: uuid }
        to_user_id: { type: string, format: uuid }
        status: { type: string }
        created_at: { type: string, format: date-time }
      required: [id, team_id, from_user_id, to_user_id, status, created_at]
security:
  - BearerAuth: []
paths:
...
  /projects/{id}/ownership-transfers:
    post:
      summary: Create project ownership transfer
      operationId: createProjectTransfer
      tags: [OwnershipTransfers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to_user_id: { type: string, format: uuid }
              required: [to_user_id]
      responses:
        "201":
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectOwnershipTransfer"
  /projects/{id}/ownership-transfers/{transfer_id}:
    delete:
      summary: Cancel project ownership transfer
      operationId: cancelProjectTransfer
      tags: [OwnershipTransfers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: transfer_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Transfer cancelled
  /ownership-transfers/projects/{transfer_id}/accept:
    post:
      summary: Accept project ownership transfer
      operationId: acceptProjectTransfer
      tags: [OwnershipTransfers]
      parameters:
        - name: transfer_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Transfer accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /ownership-transfers/projects/{transfer_id}/decline:
    post:
      summary: Decline project ownership transfer
      operationId: declineProjectTransfer
      tags: [OwnershipTransfers]
      parameters:
        - name: transfer_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Transfer declined
  /teams/{id}/ownership-transfers:
    post:
      summary: Create team ownership transfer
      operationId: createTeamTransfer
      tags: [OwnershipTransfers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to_user_id: { type: string, format: uuid }
              required: [to_user_id]
      responses:
        "201":
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamOwnershipTransfer"
  /teams/{id}/ownership-transfers/{transfer_id}:
    delete:
      summary: Cancel team ownership transfer
      operationId: cancelTeamTransfer
      tags: [OwnershipTransfers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: transfer_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Transfer cancelled
  /ownership-transfers/teams/{transfer_id}/accept:
    post:
      summary: Accept team ownership transfer
      operationId: acceptTeamTransfer
      tags: [OwnershipTransfers]
      parameters:
        - name: transfer_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Transfer accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /ownership-transfers/teams/{transfer_id}/decline:
    post:
      summary: Decline team ownership transfer
      operationId: declineTeamTransfer
      tags: [OwnershipTransfers]
      parameters:
        - name: transfer_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Transfer declined